+++ app/Http/Controllers/PedagangController.php
@@ -66,76 +66,78 @@ class PedagangController extends Controller
      */
     public function store(Request $request)
     {
-        $rules = array(
-            'ktp'      => 'required',
-            'nama'     => ['required', 'regex:/^[a-zA-Z\.\s]+$/u','min:2', 'max:30'],
-            'username' => 'required',
-            'anggota'  => 'required',
-            'hp'       => 'required',
-            'alamat'   => 'required',
-        );
+        if(request()->ajax()){
+            $rules = array(
+                'ktp'      => 'required',
+                'nama'     => ['required', 'regex:/^[a-zA-Z\.\s]+$/u','min:2', 'max:30'],
+                'username' => 'required',
+                'anggota'  => 'required',
+                'hp'       => 'required',
+                'alamat'   => 'required',
+            );
 
-        $error = Validator::make($request->all(), $rules);
+            $error = Validator::make($request->all(), $rules);
 
-        if($error->fails())
-        {
-            return response()->json(['errors' => 'Data Gagal Ditambah.']);
-        }
+            if($error->fails())
+            {
+                return response()->json(['errors' => 'Data Gagal Ditambah.']);
+            }
 
-        $data = [
-            'ktp'      => $request->ktp,
-            'nama'     => ucwords($request->nama),
-            'username' => strtolower($request->username),
-            'password' => sha1(md5(hash('gost','1234567'))),
-            'anggota'  => strtoupper($request->anggota),
-            'email'    => strtolower($request->email.'@gmail.com'),
-            'alamat'   => $request->alamat,
-            'role'     => 'nasabah',
-        ];
-       
-        if($request->email == NULL) {
-            $data['email'] = NULL;
-        }
+            $data = [
+                'ktp'      => $request->ktp,
+                'nama'     => ucwords($request->nama),
+                'username' => strtolower($request->username),
+                'password' => sha1(md5(hash('gost','1234567'))),
+                'anggota'  => strtoupper($request->anggota),
+                'email'    => strtolower($request->email.'@gmail.com'),
+                'alamat'   => $request->alamat,
+                'role'     => 'nasabah',
+            ];
         
-        if($request->hp[0] == '0') {
-            $hp = '62'.substr($request->hp,1);
-            $data['hp'] = $hp;
-        }
-        else{
-            $hp = '62'.$request->hp;
-            $data['hp'] = $hp;
-        }
-
-        try{
-            Pedagang::create($data);
-
-            $ped = Pedagang::where('username',$request->username)->first();
-
-            if($request->pemilik == "pemilik"){
-                $alamatPemilik = $request->alamatPemilik;
-                foreach($alamatPemilik as $alamat){
-                    $tempat = TempatUsaha::find($alamat);
-                    if($tempat != NULL){
-                        $tempat->id_pemilik = $ped->id;
-                        $tempat->save();
-                    }
-                }
+            if($request->email == NULL) {
+                $data['email'] = NULL;
+            }
+            
+            if($request->hp[0] == '0') {
+                $hp = '62'.substr($request->hp,1);
+                $data['hp'] = $hp;
+            }
+            else{
+                $hp = '62'.$request->hp;
+                $data['hp'] = $hp;
             }
 
-            if($request->pengguna == "pengguna"){
-                $alamatPengguna = $request->alamatPengguna;
-                foreach($alamatPengguna as $alamat){
-                    $tempat = TempatUsaha::find($alamat);
-                    if($tempat != NULL){
-                        $tempat->id_pengguna = $ped->id;
-                        $tempat->save();
+            try{
+                Pedagang::create($data);
+
+                $ped = Pedagang::where('username',$request->username)->first();
+
+                if($request->pemilik == "pemilik"){
+                    $alamatPemilik = $request->alamatPemilik;
+                    foreach($alamatPemilik as $alamat){
+                        $tempat = TempatUsaha::find($alamat);
+                        if($tempat != NULL){
+                            $tempat->id_pemilik = $ped->id;
+                            $tempat->save();
+                        }
                     }
                 }
+
+                if($request->pengguna == "pengguna"){
+                    $alamatPengguna = $request->alamatPengguna;
+                    foreach($alamatPengguna as $alamat){
+                        $tempat = TempatUsaha::find($alamat);
+                        if($tempat != NULL){
+                            $tempat->id_pengguna = $ped->id;
+                            $tempat->save();
+                        }
+                    }
+                }
+                return response()->json(['success' => 'Data Berhasil Ditambah.']);
             }
-            return response()->json(['success' => 'Data Berhasil Ditambah.']);
-        }
-        catch(\Exception $e){
-            return response()->json(['errors' => 'Data Gagal Ditambah.']);
+            catch(\Exception $e){
+                return response()->json(['errors' => 'Data Gagal Ditambah.']);
+            }
         }
     }
 
@@ -203,114 +205,116 @@ class PedagangController extends Controller
      */
     public function update(Request $request, Pedagang $pedagang)
     {
-        $rules = array(
-            'ktp'      => 'required',
-            'nama'     => ['required', 'regex:/^[a-zA-Z\.\s]+$/u','min:2', 'max:30'],
-            'hp'       => 'required',
-            'alamat'   => 'required',
-        );
+        if(request()->ajax()){
+            $rules = array(
+                'ktp'      => 'required',
+                'nama'     => ['required', 'regex:/^[a-zA-Z\.\s]+$/u','min:2', 'max:30'],
+                'hp'       => 'required',
+                'alamat'   => 'required',
+            );
 
-        $error = Validator::make($request->all(), $rules);
+            $error = Validator::make($request->all(), $rules);
 
-        if($error->fails())
-        {
-            return response()->json(['errors' => 'Data Gagal Diedit.']);
-        }
+            if($error->fails())
+            {
+                return response()->json(['errors' => 'Data Gagal Diedit.']);
+            }
 
-        $data = [
-            'ktp'      => $request->ktp,
-            'nama'     => ucwords($request->nama),
-            'email'    => strtolower($request->email.'@gmail.com'),
-            'alamat'   => $request->alamat,
-        ];
+            $data = [
+                'ktp'      => $request->ktp,
+                'nama'     => ucwords($request->nama),
+                'email'    => strtolower($request->email.'@gmail.com'),
+                'alamat'   => $request->alamat,
+            ];
 
-        if($request->email == NULL) {
-            $data['email'] = NULL;
-        }
-        
-        if($request->hp[0] == '0') {
-            $hp = '62'.substr($request->hp,1);
-            $data['hp'] = $hp;
-        }
-        else{
-            $hp = '62'.$request->hp;
-            $data['hp'] = $hp;
-        }
+            if($request->email == NULL) {
+                $data['email'] = NULL;
+            }
+            
+            if($request->hp[0] == '0') {
+                $hp = '62'.substr($request->hp,1);
+                $data['hp'] = $hp;
+            }
+            else{
+                $hp = '62'.$request->hp;
+                $data['hp'] = $hp;
+            }
 
-        try{
-            Pedagang::whereId($request->hidden_id)->update($data);
+            try{
+                Pedagang::whereId($request->hidden_id)->update($data);
 
-            $ped = Pedagang::find($request->hidden_id);
+                $ped = Pedagang::find($request->hidden_id);
 
-            if($request->pemilik == "pemilik"){
-                $alamatPemilik = $request->alamatPemilik;
+                if($request->pemilik == "pemilik"){
+                    $alamatPemilik = $request->alamatPemilik;
 
-                $tempat = TempatUsaha::where('id_pemilik',$request->hidden_id)->get();
-                if($tempat != NULL){
-                    foreach($tempat as $t){
-                        if(in_array($t->kd_kontrol, $alamatPemilik) == FALSE){
+                    $tempat = TempatUsaha::where('id_pemilik',$request->hidden_id)->get();
+                    if($tempat != NULL){
+                        foreach($tempat as $t){
+                            if(in_array($t->kd_kontrol, $alamatPemilik) == FALSE){
+                                $hapus = TempatUsaha::find($t->id);
+                                $hapus->id_pemilik = NULL;
+                                $hapus->save();
+                            }
+                        }   
+                    }
+                    
+                    foreach($alamatPemilik as $alamat){
+                        $tempat = TempatUsaha::find($alamat);
+                        if($tempat != NULL){
+                            $tempat->id_pemilik = $ped->id;
+                            $tempat->save();
+                        }
+                    }
+                }
+                else{
+                    $tempat = TempatUsaha::where('id_pemilik',$request->hidden_id)->get();
+                    if($tempat != NULL){
+                        foreach($tempat as $t){
                             $hapus = TempatUsaha::find($t->id);
                             $hapus->id_pemilik = NULL;
                             $hapus->save();
-                        }
-                    }   
-                }
-                
-                foreach($alamatPemilik as $alamat){
-                    $tempat = TempatUsaha::find($alamat);
-                    if($tempat != NULL){
-                        $tempat->id_pemilik = $ped->id;
-                        $tempat->save();
+                        }   
                     }
                 }
-            }
-            else{
-                $tempat = TempatUsaha::where('id_pemilik',$request->hidden_id)->get();
-                if($tempat != NULL){
-                    foreach($tempat as $t){
-                        $hapus = TempatUsaha::find($t->id);
-                        $hapus->id_pemilik = NULL;
-                        $hapus->save();
-                    }   
+
+                if($request->pengguna == "pengguna"){
+                    $alamatPengguna = $request->alamatPengguna;
+
+                    $tempat = TempatUsaha::where('id_pengguna',$request->hidden_id)->get();
+                    if($tempat != NULL){
+                        foreach($tempat as $t){
+                            if(in_array($t->kd_kontrol, $alamatPengguna) == FALSE){
+                                $hapus = TempatUsaha::find($t->id);
+                                $hapus->id_pengguna = NULL;
+                                $hapus->save();
+                            }
+                        }   
+                    }
+
+                    foreach($alamatPengguna as $alamat){
+                        $tempat = TempatUsaha::find($alamat);
+                        if($tempat != NULL){
+                            $tempat->id_pengguna = $ped->id;
+                            $tempat->save();
+                        }
+                    }
                 }
-            }
-
-            if($request->pengguna == "pengguna"){
-                $alamatPengguna = $request->alamatPengguna;
-
-                $tempat = TempatUsaha::where('id_pengguna',$request->hidden_id)->get();
-                if($tempat != NULL){
-                    foreach($tempat as $t){
-                        if(in_array($t->kd_kontrol, $alamatPengguna) == FALSE){
+                else{
+                    $tempat = TempatUsaha::where('id_pengguna',$request->hidden_id)->get();
+                    if($tempat != NULL){
+                        foreach($tempat as $t){
                             $hapus = TempatUsaha::find($t->id);
                             $hapus->id_pengguna = NULL;
                             $hapus->save();
-                        }
-                    }   
-                }
-
-                foreach($alamatPengguna as $alamat){
-                    $tempat = TempatUsaha::find($alamat);
-                    if($tempat != NULL){
-                        $tempat->id_pengguna = $ped->id;
-                        $tempat->save();
+                        }   
                     }
                 }
+                return response()->json(['success' => 'Data Berhasil Diedit.']);
             }
-            else{
-                $tempat = TempatUsaha::where('id_pengguna',$request->hidden_id)->get();
-                if($tempat != NULL){
-                    foreach($tempat as $t){
-                        $hapus = TempatUsaha::find($t->id);
-                        $hapus->id_pengguna = NULL;
-                        $hapus->save();
-                    }   
-                }
+            catch(\Exception $e){
+                return response()->json(['errors' => 'Data Gagal Diedit.']);
             }
-            return response()->json(['success' => 'Data Berhasil Diedit.']);
-        }
-        catch(\Exception $e){
-            return response()->json(['errors' => 'Data Gagal Diedit.']);
         }
     }
 
@@ -322,13 +326,15 @@ class PedagangController extends Controller
      */
     public function destroy($id)
     {
-        $data = Pedagang::findOrFail($id);
-        try{
-            $data->delete();
-            return response()->json(['success' => 'Data telah dihapus.']);
-        }
-        catch(\Exception $e){
-            return response()->json(['errors' => 'Data gagal dihapus.']);
+        if(request()->ajax()){
+            $data = Pedagang::find($id);
+            try{
+                $data->delete();
+                return response()->json(['success' => 'Data telah dihapus.']);
+            }
+            catch(\Exception $e){
+                return response()->json(['errors' => 'Data gagal dihapus.']);
+            }
         }
     }
 }
